# backend/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HH Scout API",
    "version": "1.0.0",
    "description": "API for HH Scout Miniapp"
  },

  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],

  "tags": [{ "name": "Jobs" }, { "name": "Favorites" }, { "name": "Areas" }, { "name": "Statuses" }],

  "paths": {
    "/jobs": {
      "get": {
        "operationId": "searchJobs",
        "tags": ["Jobs"],
        "summary": "Search jobs",
        "parameters": [
          { "name": "text", "in": "query", "schema": { "type": "string" } },

          { "name": "page", "in": "query", "schema": { "type": "number", "default": 1 } },

          { "name": "per_page", "in": "query", "schema": { "type": "number", "default": 100 } },

          { "name": "index", "in": "query", "schema": { "type": "number", "default": 0 } },

          { "name": "order_by", "in": "query", "schema": { "type": "string" } },

          { "name": "salary", "in": "query", "schema": { "type": "number" } },

          { "name": "currency", "in": "query", "schema": { "type": "string" } },

          { "name": "experience", "in": "query", "schema": { "type": "string" } },

          {
            "name": "employment_form",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          },

          {
            "name": "work_format",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          },

          {
            "name": "area",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          },

          {
            "name": "label",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          },

          { "name": "period", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": {
          "200": {
            "description": "Search result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          }
        }
      }
    },

    "/jobs/prefetch": {
      "get": {
        "operationId": "prefetchJobs",
        "tags": ["Jobs"],
        "summary": "Prefetch next jobs",
        "parameters": [
          { "name": "text", "in": "query", "schema": { "type": "string" } },

          { "name": "page", "in": "query", "schema": { "type": "number", "default": 1 } },

          { "name": "per_page", "in": "query", "schema": { "type": "number", "default": 100 } },

          { "name": "index", "in": "query", "schema": { "type": "number", "default": 0 } },

          { "name": "order_by", "in": "query", "schema": { "type": "string" } },

          { "name": "salary", "in": "query", "schema": { "type": "number" } },

          { "name": "currency", "in": "query", "schema": { "type": "string" } },

          { "name": "experience", "in": "query", "schema": { "type": "string" } },

          {
            "name": "employment_form",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          },

          {
            "name": "work_format",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          },

          {
            "name": "area",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          },

          {
            "name": "label",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          },

          { "name": "period", "in": "query", "schema": { "type": "number" } }
        ],

        "responses": {
          "200": {
            "description": "Prefetched jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/jobs/{id}": {
      "get": {
        "operationId": "getJobById",
        "tags": ["Jobs"],
        "summary": "Get job by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      }
    },

    "/areas": {
      "get": {
        "operationId": "getAreas",
        "tags": ["Areas"],
        "summary": "Get list of regions",
        "responses": {
          "200": {
            "description": "List of areas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Area"
                  }
                }
              }
            }
          }
        }
      }
    },

    "/favorites": {
      "post": {
        "operationId": "saveFavorite",
        "tags": ["Favorites"],
        "summary": "Save favorite job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FavoriteCreate" }
            }
          }
        },
        "responses": { "200": { "description": "Favorite saved" } }
      },

      "get": {
        "operationId": "getFavorites",
        "tags": ["Favorites"],
        "summary": "Get favorites with filters",
        "parameters": [
          { "name": "userId", "in": "query", "required": true, "schema": { "type": "number" } },
          { "name": "page", "in": "query", "schema": { "type": "number", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "number", "default": 20 } },
          { "name": "salary_from", "in": "query", "schema": { "type": "number" } },
          {
            "name": "company",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "status",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": { "type": "array", "items": { "type": "number" } }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["date", "salary_asc", "salary_desc", "status"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorites list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FavoritesResponse" }
              }
            }
          }
        }
      }
    },

    "/favorites/ids": {
      "get": {
        "operationId": "getFavoriteIds",
        "tags": ["Favorites"],
        "summary": "Get all favorite job ids",
        "parameters": [
          { "name": "userId", "in": "query", "required": true, "schema": { "type": "number" } }
        ],
        "responses": {
          "200": {
            "description": "List of favorite job ids",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ids": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/favorites/export": {
      "get": {
        "operationId": "exportFavoritesExcel",
        "tags": ["Favorites"],
        "summary": "Export favorites to Excel",
        "parameters": [
          { "name": "userId", "in": "query", "required": true, "schema": { "type": "number" } }
        ],
        "responses": {
          "200": {
            "description": "Excel file"
          }
        }
      }
    },

    "/favorites/{jobId}": {
      "delete": {
        "operationId": "deleteFavorite",
        "tags": ["Favorites"],
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "query", "required": true, "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "Favorite removed" } }
      }
    },

    "/favorites/{jobId}/status": {
      "patch": {
        "operationId": "setFavoriteStatus",
        "tags": ["Favorites"],
        "summary": "Change favorite status",
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId"],
                "properties": {
                  "userId": { "type": "number" },
                  "statusId": { "type": "number", "nullable": true }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Status updated" } }
      }
    },

    "/statuses": {
      "get": {
        "operationId": "getStatuses",
        "tags": ["Statuses"],
        "parameters": [
          { "name": "userId", "in": "query", "required": true, "schema": { "type": "number" } }
        ],
        "responses": {
          "200": {
            "description": "List of statuses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Status" }
                }
              }
            }
          }
        }
      },

      "post": {
        "operationId": "createStatus",
        "tags": ["Statuses"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StatusCreate" }
            }
          }
        },
        "responses": { "200": { "description": "Status created" } }
      }
    },

    "/statuses/{id}": {
      "patch": {
        "operationId": "updateStatus",
        "tags": ["Statuses"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "number" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StatusUpdate" }
            }
          }
        },
        "responses": { "200": { "description": "Status updated" } }
      },

      "delete": {
        "operationId": "deleteStatus",
        "tags": ["Statuses"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "Status deleted" } }
      }
    }
  },

  "components": {
    "schemas": {
      "Job": {
        "type": "object",
        "required": ["id", "title", "company", "url", "description"],
        "properties": {
          "id": { "type": "string" },

          "title": { "type": "string" },

          "company": { "type": "string" },

          "url": { "type": "string" },

          "description": { "type": "string" },

          "publishedAt": { "type": "string" },

          "salaryFrom": { "type": "number", "nullable": true },

          "salaryTo": { "type": "number", "nullable": true },

          "currency": { "type": "string", "nullable": true },

          "experience": { "type": "string", "nullable": true },

          "employmentForm": { "type": "string", "nullable": true },

          "workFormat": {
            "type": "array",
            "items": { "type": "string" }
          },

          "workSchedule": {
            "type": "array",
            "items": { "type": "string" }
          },

          "workingHours": {
            "type": "array",
            "items": { "type": "string" }
          },

          "city": { "type": "string", "nullable": true },

          "skills": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },

      "Favorite": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },

          "userId": { "type": "number" },

          "jobId": { "type": "string" },

          "title": { "type": "string" },

          "company": { "type": "string" },

          "salaryFrom": { "type": "number", "nullable": true },

          "salaryTo": { "type": "number", "nullable": true },

          "currency": { "type": "string", "nullable": true },

          "experience": { "type": "string", "nullable": true },

          "url": { "type": "string" }
        }
      },

      "StatusColor": {
        "type": "string",
        "enum": [
          "blue",
          "purple",
          "orange",
          "green",
          "red",
          "yellow",
          "pink",
          "teal",
          "indigo",
          "gray"
        ]
      },

      "Status": {
        "type": "object",
        "required": ["id", "userId", "name", "color"],
        "properties": {
          "id": { "type": "number" },
          "userId": { "type": "number" },
          "name": { "type": "string" },
          "color": { "$ref": "#/components/schemas/StatusColor" }
        }
      },

      "StatusCreate": {
        "type": "object",
        "required": ["userId", "name", "color"],
        "properties": {
          "userId": { "type": "number" },
          "name": { "type": "string" },
          "color": { "$ref": "#/components/schemas/StatusColor" }
        }
      },

      "StatusUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "color": { "$ref": "#/components/schemas/StatusColor" }
        }
      },

      "FavoritesResponse": {
        "type": "object",
        "required": ["items", "meta"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Favorite" }
          },
          "meta": {
            "type": "object",
            "properties": {
              "total": { "type": "number" },
              "page": { "type": "number" },
              "pages": { "type": "number" }
            }
          },
          "filters": {
            "type": "object",
            "properties": {
              "companies": {
                "type": "array",
                "items": { "type": "string" }
              },
              "statuses": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Status" }
              }
            }
          }
        }
      },

      "FavoriteCreate": {
        "type": "object",
        "required": ["userId", "job"],
        "properties": {
          "userId": { "type": "number" },

          "job": {
            "$ref": "#/components/schemas/Job"
          }
        }
      },

      "SearchResult": {
        "type": "object",
        "required": ["items", "page", "pages", "found", "perPage", "pageItems"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            }
          },

          "page": { "type": "number" },

          "pages": { "type": "number" },

          "found": { "type": "number" },

          "perPage": { "type": "number" },

          "pageItems": { "type": "number" }
        }
      },

      "Area": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" }
        }
      }
    }
  }
}